refactor virtual wards
previously we defined all of the logic in activity avoidance, then imported activity avoidance into efficiencies.

however, that could cause an error as the activity avoidance mitigators would try to be defined twice when run with the reflective import (in raw_data.inpatients_mitigators.py).

moved the logic into a new 'shared' folder where we can re-use logic as needed