fixes birth extract logic
there exists an issue with the birth factor logic where it needs to read the files that it's about to save. this works if you have previously run the extract, but fails if you are loading to a new location.

this refactors the logic so we first create the standard birth factors and save to disk, before running the custom factors after.